Split mypy.myunit package into modules
Part of #908
