+++ mypy/myunit/errors.py
@@ -1,0 +1,7 @@
+class MyunitException(Exception): pass
+
+
+class AssertionFailure(MyunitException): pass
+
+
+class SkipTestCaseException(MyunitException): pass
