+++ mypy/myunit/ick.py
@@ -1,0 +1,6 @@
+"""Ick, mutable globals!
+"""
+
+# TODO remove global state
+APPEND_TESTCASES = ''
+UPDATE_TESTCASES = False
