+++ mypy/test/testsemanal.py
@@ -5,7 +5,7 @@ import os.path
 from typing import Dict, List
 
 from mypy import build
-from mypy.myunit import Suite
+from mypy.myunit.suite import Suite
 from mypy.test.helpers import assert_string_arrays_equal, testfile_pyversion
 from mypy.test.data import parse_test_cases
 from mypy.test.config import test_data_prefix, test_temp_dir
